smart_kiosk/requirements.txt
Flask
flask-cors
python-dotenv
pandas
openpyxl
requests
langchain
langchain-community
langchain-huggingface
langchain-text-splitters
langchain-chroma
chromadb
pypdf
docx2txt
sentence-transformers
